use tape-promise, not baseline tape, for tests that return promises
Our unit tests use async/await, so they return Promises. 'tape' doesn't pay
attention to the return value of a test case, so it will drop that promise on
the floor. I think our tests happened to work anyways because we got lucky,
but we really ought to be using a runner like tape-promise that waits for the
promise to resolve before considering the test complete.